video-call-microservice/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
Flask-SocketIO==5.3.6